Generates helper code after creating a new eloquent model.
