+++ src/main/java/at/alirezamoh/whisperer_for_laravel/actions/EloquentAction.java
@@ -6,7 +6,10 @@ import at.alirezamoh.whisperer_for_laravel.actions.models.MigrationModel;
 import at.alirezamoh.whisperer_for_laravel.actions.models.EloquentModel;
 import at.alirezamoh.whisperer_for_laravel.actions.views.EloquentView;
 import at.alirezamoh.whisperer_for_laravel.settings.SettingsState;
+import at.alirezamoh.whisperer_for_laravel.support.codeGeneration.HelperCodeExecutor;
 import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.progress.EmptyProgressIndicator;
 import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -43,9 +46,18 @@ public class EloquentAction extends BaseAction {
             if (dbFactoryModel != null) {
                 createFactory(dbFactoryModel, project);
             }
+
+            generateHelperCode(project);
         }
     }
 
+    /**
+     * Creates a controller based on the provided EloquentView and EloquentModel
+     *
+     * @param eloquentView   The view containing user input for the controller
+     * @param eloquentModel  The model containing data for the controller
+     * @param project        The current project
+     */
     private void createController(EloquentView eloquentView, EloquentModel eloquentModel, Project project) {
         ControllerModel controllerModel = new ControllerModel(
             SettingsState.getInstance(project),
@@ -61,6 +73,13 @@ public class EloquentAction extends BaseAction {
         );
     }
 
+    /**
+     * Creates a migration file based on the provided EloquentView and EloquentModel
+     *
+     * @param eloquentView   The view containing user input for the migration
+     * @param eloquentModel  The model containing data for the migration
+     * @param project        The current project
+     */
     private void createMigration(EloquentView eloquentView, EloquentModel eloquentModel, Project project) {
         MigrationModel migrationModel = new MigrationModel(
             SettingsState.getInstance(project),
@@ -82,6 +101,12 @@ public class EloquentAction extends BaseAction {
         );
     }
 
+    /**
+     * Creates a database factory based on the provided DbFactoryModel
+     *
+     * @param factoryModel  The model containing data for the factory
+     * @param project       The current project
+     */
     private void createFactory(DbFactoryModel factoryModel, Project project) {
         this.create(
             factoryModel,
@@ -91,6 +116,14 @@ public class EloquentAction extends BaseAction {
         );
     }
 
+    /**
+     * Creates a DbFactoryModel if the EloquentView indicates that a factory should be created
+     *
+     * @param eloquentView   The view containing user input for the factory
+     * @param eloquentModel  The model containing data for the factory
+     * @param project        The current project
+     * @return A DbFactoryModel if the factory should be created, null otherwise
+     */
     private @Nullable DbFactoryModel createDbFactory(EloquentView eloquentView, EloquentModel eloquentModel, Project project) {
         if (eloquentView.withFactory()) {
             DbFactoryModel dbFactoryModel = new DbFactoryModel(
@@ -107,5 +140,22 @@ public class EloquentAction extends BaseAction {
         }
 
         return null;
+    }
+
+    /**
+     * Generates helper code in the background
+     *
+     * @param project The current project context
+     */
+    private void generateHelperCode(Project project) {
+        ApplicationManager.getApplication().executeOnPooledThread(() -> {
+            HelperCodeExecutor helperCodeExecutor = new HelperCodeExecutor(
+                project,
+                new EmptyProgressIndicator(),
+                false
+            );
+
+            helperCodeExecutor.execute();
+        });
     }
 }
